+++ keba_keenergy_api/constants.py
@@ -792,6 +792,10 @@ class HeatCircuit(Enum):
         f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.heatCurveOffset",
         value_type=float,
     )
+    COOLING_CURVE_OFFSET = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.coolCurveOffset",
+        value_type=float,
+    )
     HEATING_CURVE_SLOPE = EndpointProperties(
         f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.heatCurveGradient",
         value_type=float,