+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `set_cooling_curve()` to heating circuit
 - Added endpoint `get_cooling_curve_slope()` to heating circuit
 - Added endpoint `set_cooling_curve_slope()` to heating circuit
+- Added endpoint `get_cooling_curve_offset()` to heating circuit
+- Added endpoint `set_cooling_curve_offset()` to heating circuit
 
 ## [2.10.2] - 2026-04-03